% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{affiliation_dates}
\alias{affiliation_dates}
\title{Affiliation dates}
\format{
A data frame with 4 variables:
\describe{
\item{\code{Member.ID}}{The ID number of the individual.}
\item{\code{Org.ID}}{The ID number of the organization}
\item{\code{Start.Date}}{The date the individual first became involved with the organization.}
\item{\code{End.Date}}{The date the individual left the organization; either by de-affiliating, or because the event or institude ended.}
}
}
\usage{
affiliation_dates
}
\description{
Table with start and end dates for each member's affiliation to various organizations.
}
\keyword{datasets}
