% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_metrics_by_month}
\alias{all_metrics_by_month}
\title{All pre-calculated metrics}
\format{
A data frame with N variables:
\describe{
\item{\code{Member.ID}}{The ID number of the individual.}
\item{\code{Start.Date}}{The date the individual first became involved with the organization.}
\item{\code{End.Date}}{The date the individual left the organization; either by de-affiliating, or because the event or institude ended.}
}
}
\usage{
all_metrics_by_month
}
\description{
Table all metrics calculated for each individual for each month
}
\keyword{datasets}
