% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_metrics_org.R
\name{betweenness_checked_orgs}
\alias{betweenness_checked_orgs}
\title{Calculates betweenness but first checks if vertex is valid}
\usage{
betweenness_checked_orgs(graph, edgelist, vertex)
}
\arguments{
\item{graph}{An igraph object}

\item{edgelist}{A tibble of edges}

\item{vertex}{The vertex to calculate betweenness of}
}
\value{
A double
}
\description{
Calculates betweenness but first checks if vertex is valid
}
