% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_metrics_mem.R, R/get_degree.R
\name{degree_checked}
\alias{degree_checked}
\title{Calculates betweenness but first checks if vertex is valid}
\usage{
degree_checked(graph, edgelist, vertex)

degree_checked(graph, edgelist, vertex)
}
\arguments{
\item{graph}{An igraph object}

\item{edgelist}{A tibble of edges}

\item{vertex}{The vertex to calculate degree of}
}
\value{
A double

A double
}
\description{
Calculates betweenness but first checks if vertex is valid

Calculates degree but first checks if vertex is valid
}
