% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_degree_weighted.R
\name{degree_weighted_checked}
\alias{degree_weighted_checked}
\title{Calculates weighted degree but first checks if vertex is valid}
\usage{
degree_weighted_checked(graph, edgelist, vertex)
}
\arguments{
\item{graph}{An igraph object}

\item{edgelist}{A tibble of edges}

\item{vertex}{The vertex to calculate weighted degree of}
}
\value{
A double
}
\description{
Calculates weighted degree but first checks if vertex is valid
}
