% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_metrics_org.R
\name{get_all_metrics_orgs}
\alias{get_all_metrics_orgs}
\title{Find the degree, betweenness, cross-group degree, and normalized versions of
those measures for a set of Organizations over a time period}
\usage{
get_all_metrics_orgs(
  affils_by_date,
  weight_by,
  totals,
  min_cons = 1,
  orgs = NULL,
  start = NULL,
  end = NULL,
  timesteps = "months"
)
}
\arguments{
\item{affils_by_date}{The data frame containing edge connections.  Format must match
built-in dataset \code{affiliation_dates}}

\item{weight_by}{Which option to compute edge weights}

\item{min_cons}{min connections to keep an edge}

\item{orgs}{A character vector of org IDs to limit computations to}

\item{start}{A start of date range, in DD-MM-YYYY string format.  Defaults to beginning of dates in data.}

\item{end}{An end of date range, in DD-MM-YYYY string format. Defaults to end of dates in data.}

\item{timesteps}{A string ("days", "months", "years") for the level of aggregation
of the affils_by_date before computing metrics}
}
\value{
A tibble of individual Member IDs and their metrics
}
\description{
Find the degree, betweenness, cross-group degree, and normalized versions of
those measures for a set of Organizations over a time period
}
