% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_network.R
\name{get_cons_by_afil_orgs}
\alias{get_cons_by_afil_orgs}
\title{Helper Function for get_edgelist_orgs to get total member connections by affiliations}
\usage{
get_cons_by_afil_orgs(affils_by_date, affil_by, start, end = NULL)
}
\arguments{
\item{affils_by_date}{Data frame of affiliations with start and end dates}

\item{affil_by}{which affiliation to get stats for}

\item{start}{A start of date range, in YYYY-MM-DD string format.}

\item{end}{An end of date range, in YYYY-MM-DD string format.}
}
\description{
Helper Function for get_edgelist_orgs to get total member connections by affiliations
}
