% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_degree_weighted.R
\name{get_degree_weighted_orgs}
\alias{get_degree_weighted_orgs}
\title{Find the weighted degree measures for a set of organizations over a time period}
\usage{
get_degree_weighted_orgs(orgs, start = NULL, end = NULL, timesteps = "months")
}
\arguments{
\item{orgs}{A character vector of organization IDs.}

\item{start}{A start of date range, in DD-MM-YYYY string format.  Defaults to beginning of dates in data.}

\item{end}{An end of date range, in DD-MM-YYYY string format. Defaults to end of dates in data.}
}
\value{
A tibble with pairs of organizations and their number of shared members in that range.
}
\description{
Find the weighted degree measures for a set of organizations over a time period
}
