% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_network.R
\name{get_edgelist_members}
\alias{get_edgelist_members}
\title{Make an edgelist of members with organizational overlap in given date range}
\usage{
get_edgelist_members(start, end = NULL)
}
\arguments{
\item{start}{A start of date range, in DD-MM-YYYY string format.}

\item{end}{An end of date range, in DD-MM-YYYY string format.}
}
\value{
A tibble with pairs of members who were in the same institution or participated in the same event in the date range.
}
\description{
Make an edgelist of members with organizational overlap in given date range
}
