% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_metrics_org.R
\name{get_one_metrics_step_org}
\alias{get_one_metrics_step_org}
\title{Find the betweenness measures for one individual in a time period}
\usage{
get_one_metrics_step_org(
  affils_by_date,
  weight_by,
  totals,
  min_cons = 1,
  orgs = NULL,
  start = NULL,
  end = NULL
)
}
\arguments{
\item{affils_by_date}{The data frame containing edge connections.  Format must match
built-in datasets \code{affiliation_dates}}

\item{weight_by}{Which columns to use to compute edges}

\item{min_cons}{Weight column}

\item{orgs}{A character vector of org IDs.}

\item{start}{A start of date range, in DD-MM-YYYY string format.  Defaults to beginning of dates in data.}

\item{end}{An end of date range, in DD-MM-YYYY string format. Defaults to end of dates in data.}

\item{timesteps}{A string ("days", "months", "years") for the level of aggregation
of the affils_by_date before computing betweenness.}
}
\value{
A tibble with the metrics of all invididuals in one time window.
}
\description{
Find the betweenness measures for one individual in a time period
}
