% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vertex_metrics.R
\name{get_one_vertex_metrics_orgs}
\alias{get_one_vertex_metrics_orgs}
\title{Find the vertex_metrics measures for a set of organizations over a time period}
\usage{
get_one_vertex_metrics_orgs(orgs, metrics, start = NULL, end = NULL)
}
\arguments{
\item{start}{A start of date range, in DD-MM-YYYY string format.  Defaults to beginning of dates in data.}

\item{end}{An end of date range, in DD-MM-YYYY string format. Defaults to end of dates in data.}

\item{...}{A set of organization IDs.}
}
\value{
A tibble with pairs of organizations and their number of shared members in that range.
}
\description{
Find the vertex_metrics measures for a set of organizations over a time period
}
