% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shared.R
\name{get_shared_members}
\alias{get_shared_members}
\title{Find membership overlaps between two organizations}
\usage{
get_shared_members(org_1, org_2)
}
\arguments{
\item{org_1}{An ID for an organization}

\item{org_2}{An ID for an organization}
}
\value{
A tibble with all member overlaps and the date range of the overlap
}
\description{
Find membership overlaps between two organizations
}
