% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shared.R
\name{get_shared_orgs}
\alias{get_shared_orgs}
\title{Find organizational overlaps between two individuals}
\usage{
get_shared_orgs(member_1, member_2)
}
\arguments{
\item{member_1}{An ID for a Round Table Member}

\item{member_2}{An ID for a Round Table Member}
}
\value{
A tibble with all organizational overlaps and the date range of the overlap
}
\description{
Find organizational overlaps between two individuals
}
