% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vertex_metrics.R
\name{get_vertex_metrics_orgs}
\alias{get_vertex_metrics_orgs}
\title{Find the vertex_metrics measures for a set of organizations over a time period}
\usage{
get_vertex_metrics_orgs(
  orgs,
  metrics,
  start = NULL,
  end = NULL,
  timesteps = "months"
)
}
\arguments{
\item{orgs}{A character vector of organization IDs.}

\item{metrics}{A named list of metric computing functions, as from \code{igraph}.  Must take arguments of (\code{graph}, \code{vertex}).}

\item{start}{A start of date range, in DD-MM-YYYY string format.  Defaults to beginning of dates in data.}

\item{end}{An end of date range, in DD-MM-YYYY string format. Defaults to end of dates in data.}
}
\value{
A tibble with pairs of organizations and their number of shared members in that range.
}
\description{
Find the vertex_metrics measures for a set of organizations over a time period
}
