% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{member_meta_info}
\alias{member_meta_info}
\title{Additional information about members of the Polish Round Table opposition.}
\format{
A data frame with 6 variables:
\describe{
\item{\code{Member.ID}}{ID number used for network matching}
\item{\code{Full.Name}}{Name of individual}
\item{\code{Age}}{Age at time of Round Table.}
\item{\code{Gender}}{Identified gender at time of Round Table}
\item{\code{RT.Affiliation}}{Currently only "Opposition"}
\item{\code{Profession}}{Category of profession at time of Round Table}
}
}
\usage{
member_meta_info
}
\description{
Additional information about members of the Polish Round Table opposition.
}
\keyword{datasets}
