% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{organization_meta_info}
\alias{organization_meta_info}
\title{Additional information about organizations (institutions and/or events).}
\format{
A data frame with 7 variables:
\describe{
\item{\code{Org.ID}}{ID code used for network matching}
\item{\code{Org.Name}}{Long name for the institute or event.}
\item{\code{Org.Group.1}}{Short identifier for the institute or event; for example, "WRPOLY" for Wroclaw Polytechnic University}
\item{\code{Org.Group.2}}{Short identifier for a subcategory of the institute or event; for example, "FMIN" for "Faculty of Mining"}
\item{\code{Type}}{Category of organization.}
\item{\code{City}}{Primary city location of institute or event.}
#' \item{\code{Vovoidship}}{Primary regional location of institute or event.}
}
}
\usage{
organization_meta_info
}
\description{
Additional information about organizations (institutions and/or events).
}
\keyword{datasets}
