% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_plots.R
\name{plot_metric}
\alias{plot_metric}
\title{Find the betweenness measures for a set of individuals over a time period}
\usage{
plot_metric(
  metric_df,
  metric_col,
  group_col = Member.ID,
  rolling_avg = 6,
  start = NULL,
  end = NULL
)
}
\arguments{
\item{metric_df}{A dataframe of betweenness measures by date, in the
format of built-in dataset \code{all_metrics}}

\item{metric_col}{An unquoted variable name specifying a column with the metric
of interest.}

\item{group_col}{(Optional) An unquoted variable name specifying a column to group by.}

\item{rolling_avg}{An integer.  Number of months before and after target
date to compute a rolling average of, for smoothing.  Use \code{0} for no smoothing.}

\item{start}{A start of date range, in DD-MM-YYYY string format.  Defaults to beginning of dates in data.}

\item{end}{An end of date range, in DD-MM-YYYY string format. Defaults to end of dates in data.}
}
\value{
A ggplot object showing the metric of individuals over time.
}
\description{
Find the betweenness measures for a set of individuals over a time period
}
