% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vertex_metrics.R
\name{vertex_metrics_checked}
\alias{vertex_metrics_checked}
\title{Calculates vertex_metrics but first checks if vertex is valid}
\usage{
vertex_metrics_checked(graph, metrics, edgelist, vertex)
}
\arguments{
\item{graph}{An igraph object}

\item{metrics}{A named list of metric functions}

\item{edgelist}{A tibble of edges}

\item{vertex}{The vertex to calculate vertex_metrics of}
}
\value{
A double
}
\description{
Calculates vertex_metrics but first checks if vertex is valid
}
